+++ packages/client/src/components/index.ts
@@ -1,5 +1,5 @@
 import App from './app'
 import TextField from './text-field'
-import Form from './from'
+import Form from './form'
 
 export { App, TextField, Form }
